Cleaned Code, implemented tiles
Removed buildAxes() method/construct, as well as initial walls

Added makeFloorPlan() which returns a 2D array of booleans, used in the generateRoom() method for constructing walls. 
Added a grid pattern to the floor so movement/cells can be easily seen.
 
  GenerateRoom() uses the 2D array of booleans to construct walls which are 1-cell wide/long. This creates awkward light patterns at the edges of the walls depending on viewing angle.

Number of tiles and tile size are now global variables.