Zombie Movement/Start Screen added
Zombies are now spawned across the room based on the inputted values. They are all currently random walk zombies since collision checking is necessary for line-walk zombies to not disappear from view in the first few seconds of the game's start.

Start Screen implemented. Required moving the construction of the main scene outside of the start() method to prevent resizing once the player got past the start screen. 

Superfluous code cleaned up, most notably in the handleMouse() method, in which I deleted a handler which was originally purposed to allow the user to continue to rotate their field of view once they hit the edge of the screen with the mouse.