Additional Constructor values
The Zombie class now natively holds its speed and smell distance.  In the future, it can also hold its A* algorithm as long as we pass it in the floorPlan of the house.

Added an extra XForm so that movement more closely reflects the movement of the camera in the main class.